# Optimize delete children