--- a/.clj-kondo/config/modules/config.edn
+++ b/.clj-kondo/config/modules/config.edn
@@ -45,6 +45,7 @@
            events
            legacy-mbql
            lib
+           lib-be
            model-persistence
            models
            permissions