--- a/src/metabase/actions/foreign_keys.clj
+++ b/src/metabase/actions/foreign_keys.clj
@@ -33,11 +33,11 @@
           (update-in [:results item-type] (fnil into #{}) new-items)))))
 
 (defn- step
-  [metadata children-fn state]
+  [metadata-fn children-fn state]
   (let [[[parent-type items] state'] (pop-queue state)]
     (if-not parent-type
       state'
-      (let [type-metadata (get metadata parent-type)
+      (let [type-metadata (metadata-fn parent-type)
             child-keys    (for [relationship type-metadata]
                             (children-fn relationship items))]
         (reduce (fn [state [item-type items]]
@@ -52,19 +52,19 @@
    :items     results})
 
 (defn- walk*
-  [item-type items metadata children-fn {:keys [max-queries]
-                                         :or   {max-queries 100}}]
+  [item-type items metadata-fn children-fn {:keys [max-queries]
+                                            :or   {max-queries 100}}]
   (reduce
    (fn [state _]
-     (step metadata children-fn state))
+     (step metadata-fn children-fn state))
    {:queue  (conj PersistentQueue/EMPTY [item-type items])
     :results {item-type (set items)}}
    (range max-queries)))
 
 (defn walk
   "Given some starting items, return their descendants."
-  [item-type items metadata children-fn & {:as opts}]
-  (-> (walk* item-type items metadata children-fn opts)
+  [item-type items metadata-fn children-fn & {:as opts}]
+  (-> (walk* item-type items metadata-fn children-fn opts)
       (update :results (fn [results]
                          (u/remove-nils
                           (update results item-type (comp not-empty set/difference) items))))
@@ -72,15 +72,15 @@
 
 (defn count-descendants
   "Given some starting items, count the number of descendants they have, according to their types."
-  [item-type items metadata children-fn & {:as opts}]
-  (let [{:keys [complete? items]} (walk item-type items metadata children-fn opts)]
+  [item-type items metadata-fn children-fn & {:as opts}]
+  (let [{:keys [complete? items]} (walk item-type items metadata-fn children-fn opts)]
     {:complete? complete?
      :counts    (update-vals items count)}))
 
 (defn delete-recursively
   "Delete the given items, along with all their descendants."
-  [item-type items metadata children-fn delete-fn & {:as opts}]
-  (let [{:keys [queue results]} (walk* item-type items metadata children-fn opts)]
+  [item-type items metadata-fn children-fn delete-fn & {:as opts}]
+  (let [{:keys [queue results]} (walk* item-type items metadata-fn children-fn opts)]
     (if (seq queue)
       (throw (ex-info "Cannot delete all descendants, as we could not enumerate them" {:queue queue}))
       (do (delete-fn results)