--- a/src/metabase/driver/sql_jdbc/actions.clj
+++ b/src/metabase/driver/sql_jdbc/actions.clj
@@ -771,23 +771,23 @@
                                   field-name->id (table-id->pk-field-name->id (:id database) table-id)]
                               [table-id (keys field-name->id)]))
         delete-children?  (:delete-children actions/*params*)
-        [errors results] (batch-execution-by-table-id!
-                          {:inputs        inputs
-                           :row-action    :model.row/delete
-                           :row-fn        (if delete-children? row-delete!*-with-children row-delete!*)
-                           :validate-fn   (fn [database table-id rows]
-                                            (let [pk-name->id (table-id->pk-field-name->id (:id database) table-id)]
-                                              (check-consistent-row-keys rows)
-                                              (check-rows-have-expected-columns-and-no-other-keys rows (keys pk-name->id))
-                                              (check-unique-rows rows)
-                                              (when-not delete-children?
-                                                (check-rows-have-no-children (:id database) table-id rows))))
-                           :input-fn      (fn [{db-id :id} table-id row]
-                                            {:database db-id
-                                             :type     :query
-                                             :query    {:source-table table-id
-                                                        :filter       (row->mbql-filter-clause
-                                                                       (table-id->pk-field-name->id db-id table-id) row)}})})]
+        [errors results]  (batch-execution-by-table-id!
+                           {:inputs        inputs
+                            :row-action    :model.row/delete
+                            :row-fn        (if delete-children? row-delete!*-with-children row-delete!*)
+                            :validate-fn   (fn [database table-id rows]
+                                             (let [pk-name->id (table-id->pk-field-name->id (:id database) table-id)]
+                                               (check-consistent-row-keys rows)
+                                               (check-rows-have-expected-columns-and-no-other-keys rows (keys pk-name->id))
+                                               (check-unique-rows rows)
+                                               (when-not delete-children?
+                                                 (check-rows-have-no-children (:id database) table-id rows))))
+                            :input-fn      (fn [{db-id :id} table-id row]
+                                             {:database db-id
+                                              :type     :query
+                                              :query    {:source-table table-id
+                                                         :filter       (row->mbql-filter-clause
+                                                                        (table-id->pk-field-name->id db-id table-id) row)}})})]
     (when (seq errors)
       (throw (ex-info (tru "Error(s) deleting rows.")
                       {:status-code 400