- Optimized 
  - delete 1 row:  126ms (62 DB calls) vs 84ms (41 DB calls)
  - delete 3 rows: 294ms (137 DB calls) vs 90ms (55 DB calls)
  - I'm pretty sure others actions are benefited as well as one of the optimization was to cache metadata provider
    - update reduced from 51 to 39 calls for a single update
- Changed to check row descendants on normal deletion => even deleting rows from table with cascade delete will also be detected ( it was ignored before )
- New error response:
```json
{
  "errors": {
    "type": "metabase.actions.error/children-exist",
    "children-count": {
      "2": 98
    }
  }
}
```